requirements.txt
absl-py==0.15.0
astunparse==1.6.3
blobfile==2.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.4
click==8.1.3
clip==0.2.0
cryptography==37.0.1
cupy-cuda111==11.5.0
cycler==0.11.0
diffusers==0.14.0
docker-pycreds==0.4.0
dominate==2.7.0
einops==0.4.1
et-xmlfile==1.1.0
fastrlock==0.8.1
filelock==3.10.0
flatbuffers==1.12
fonttools==4.33.3
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.13.3
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.12.0
importlib-resources==5.12.0
joblib==1.1.0
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
kornia==0.6.6
lxml==4.9.2
Markdown==3.4.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpi4py==3.1.4
natsort==8.3.0
networkx==2.8.4
neurite==0.2
nibabel==5.1.0
numpy==1.23.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-nvjitlink-cu12==12.3.101
oauthlib==3.2.0
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.0
pathtools==0.1.2
Pillow==9.0.1
pip==21.2.4
piq==0.7.1
prefetch-generator==1.0.3
promise==2.3
protobuf==3.19.4
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pystrum==0.3
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.3.0
PyYAML==6.0
regex==2023.3.22
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.19.3
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.12.0
sentry-sdk==1.7.2
setproctitle==1.2.3
setuptools==61.2.0
Shapely==1.8.2
shortuuid==1.0.9
six==1.15.0
smmap==5.0.0
ssim==0.3.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.5.1
tensorflow-estimator==2.5.0
termcolor==1.1.0
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tifffile==2022.5.4
timm==1.0.15
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
typing-extensions==3.7.4.3
urllib3==1.26.9
utils==1.0.1
voxelmorph==0.2
wandb==0.12.21
Werkzeug==2.1.2
wheel==0.37.1
wrapt==1.12.1
zipp==3.8.1
